refactore(document): rename documents
